修改spring事物管理，捕获所有异常，保持事物一致性

+++ src/main/java/com/wei/service/impl/DeptServiceImpl.java
@@ -3,6 +3,8 @@ package com.wei.service.impl;
 import com.wei.mapper.DeptMapper;
 import com.wei.mapper.EmpMapper;
 import com.wei.pojo.Dept;
+import com.wei.pojo.DeptLog;
+import com.wei.service.DeptLogService;
 import com.wei.service.DeptService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,6 +25,9 @@ public class DeptServiceImpl implements DeptService {
     @Autowired
     private EmpMapper empMapper;
 
+    @Autowired
+    private DeptLogService deptLogService;
+
     //查询所有部门数据
     @Override
     public List<Dept> list() {
@@ -31,12 +36,25 @@ public class DeptServiceImpl implements DeptService {
     }
 
     //根据id删除部门
-    @Transactional  //spring事物管理，保持事物一致性
+    @Transactional(rollbackFor = Exception.class)  //spring事物管理，保持事物一致性(默认只捕获运行时异常)
     @Override
-    public void delete(Integer id) {
-        deptMapper.deleteById(id);
-        //同时删除部门下的所有员工
-        empMapper.deleteByDeptId(id);
+    public void delete(Integer id) throws Exception{
+        try {
+            deptMapper.deleteById(id);
+            //模拟异常
+            if (true) {
+                throw new Exception("出现异常了~~~");
+            }
+            //同时删除部门下的所有员工
+            empMapper.deleteByDeptId(id);
+        } finally {
+            //不论是否有异常，最终都要执行的代码：记录日志
+            DeptLog deptLog = new DeptLog();
+            deptLog.setCreateTime(LocalDateTime.now());
+            deptLog.setDescription("执行了解散部门的操作，此时解散的是" + id + "号部门");
+            //调用其他业务类中的方法
+            deptLogService.insert(deptLog);
+        }
     }
 
     //添加部门
